Implement insert head node
Duplicate the input string to a variable.
Create an element node, then set the value of the node.
After initialization, insert the node at the beginning of the list.
If one of the creations, value, or element fails,
free another variable and return false.
If there is no head node, return false.

Change-Id: Icb4bfb169562c224e3aa712880a6004a17b9dd3a