Implement insert tail node
Duplicate the input string to a variable.
Create an element node, then set the value of the node.
After initialization, insert the node at the beginning of the list.
If insertion is successful, return true.
If one of the creations, value, or element fails,
free another variable and return false.
If there is no head node, return false.

Change-Id: I1a61bce3ff3f6060ceaa604696eba7821505d9c5