Implement merge function
Find the minimum or maximum value
from the first node of all sub-queues.
Whenever the minimum or maximum value is found,
remove the node from the sub-queue
and add it to the tail of a temporary queue.

Use two loops to traverse the nodes.
One for traversing all sub-queues,
and another for checking all nodes.

If there are no header node or no queues in the chain,
return zero.
If there is only one queue in the chain,
return the size of the queue.

Change-Id: Idee565658b6845a33b645103e5ea2d8474f9cbda